refactor : [ #none ddd 구조로 변경 ]
