+++ MemberService/src/main/java/com/memberservice/member/dto/request/RequestLoginDto.java
@@ -1,8 +1,10 @@
-package com.memberservice.dto.request;
+package com.memberservice.member.dto.request;
 
 import lombok.Getter;
+import lombok.ToString;
 
 @Getter
+@ToString
 public class RequestLoginDto {
 
     private String email;
