+++ MemberService/src/main/java/com/memberservice/member/dto/request/SignUpMemberDto.java
@@ -1,10 +1,9 @@
-package com.memberservice.dto.request;
+package com.memberservice.member.dto.request;
 
-import com.memberservice.entity.Gender;
-import com.memberservice.entity.Role;
+import com.memberservice.member.entity.Gender;
+import com.memberservice.member.entity.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
