+++ MemberService/src/main/java/com/memberservice/member/dto/response/ResponseLoginDto.java
@@ -1,4 +1,4 @@
-package com.memberservice.dto.response;
+package com.memberservice.member.dto.response;
 
 public class ResponseLoginDto {
